requirements.txt
tensorflow
keras
numpy
kaggle
pillow
fastapi
streamlit
uvicorn
multipart